refactor: Role 도메인 삭제, enum으로 변경
