--- api/api-member/src/main/java/com/pgms/apimember/dto/request/PermissionUpdateRequest.java
@@ -1,11 +1,0 @@
-package com.pgms.apimember.dto.request;
-
-import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.Size;
-
-public record PermissionUpdateRequest(
-	@NotBlank(message = "이름을 입력해주세요.")
-	@Size(min = 2, max = 20, message = "이름은 2자 이상 20자 이하로 입력해주세요.")
-	String name
-) {
-}
