--- api/api-member/src/main/java/com/pgms/apimember/dto/response/RoleGetResponse.java
@@ -1,16 +1,0 @@
-package com.pgms.apimember.dto.response;
-
-import java.time.LocalDateTime;
-
-import com.pgms.coredomain.domain.member.Role;
-
-public record RoleGetResponse(
-	Long id,
-	String name,
-	LocalDateTime createdAt,
-	LocalDateTime updatedAt
-) {
-	public static RoleGetResponse from(Role role) {
-		return new RoleGetResponse(role.getId(), role.getName(), role.getCreatedAt(), role.getUpdatedAt());
-	}
-}
