--- core/core-domain/src/main/java/com/pgms/coredomain/domain/member/Permission.java
@@ -1,36 +1,0 @@
-package com.pgms.coredomain.domain.member;
-
-import com.pgms.coredomain.domain.common.BaseEntity;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.Table;
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Entity
-@Table(name = "permission")
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class Permission extends BaseEntity {
-
-	@Id
-	@Column(name = "id", nullable = false)
-	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Long id;
-
-	@Column(name = "name", nullable = false)
-	private String name;
-
-	public Permission(String name) {
-		this.name = name;
-	}
-
-	public void updateName(String name) {
-		this.name = name;
-	}
-}
