--- core/core-domain/src/main/java/com/pgms/coredomain/domain/member/Role.java
@@ -1,54 +1,0 @@
-package com.pgms.coredomain.domain.member;
-
-import java.util.ArrayList;
-import java.util.List;
-
-import com.pgms.coredomain.domain.common.BaseEntity;
-
-import jakarta.persistence.CascadeType;
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.OneToMany;
-import jakarta.persistence.Table;
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Entity
-@Table(name = "role")
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class Role extends BaseEntity {
-
-	@Id
-	@Column(name = "id", nullable = false)
-	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Long id;
-
-	@Column(name = "name", nullable = false)
-	private String name;
-
-	@OneToMany(mappedBy = "role", cascade = CascadeType.REMOVE)
-	private List<RolePermission> rolePermissions = new ArrayList<>();
-
-	public Role(String name) {
-		this.name = name;
-	}
-
-	public void updateName(String name) {
-		this.name = name;
-	}
-
-	public void addPermissionToRole(RolePermission rolePermission) {
-		if (!rolePermissions.contains(rolePermission)) {
-			rolePermissions.add(rolePermission);
-		}
-	}
-
-	public void removePermissionFromRole(RolePermission rolePermission) {
-		rolePermissions.remove(rolePermission);
-	}
-}
