--- core/core-domain/src/main/java/com/pgms/coredomain/domain/member/RolePermission.java
@@ -1,39 +1,0 @@
-package com.pgms.coredomain.domain.member;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.FetchType;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
-import jakarta.persistence.Table;
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Entity
-@Table(name = "role_permission")
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class RolePermission {
-
-	@Id
-	@Column(name = "id", nullable = false)
-	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Long id;
-
-	@ManyToOne(fetch = FetchType.LAZY, optional = false)
-	@JoinColumn(name = "role_id", nullable = false)
-	private Role role;
-
-	@ManyToOne(fetch = FetchType.LAZY, optional = false)
-	@JoinColumn(name = "permission_id", nullable = false)
-	private Permission permission;
-
-	public RolePermission(Role role, Permission permission) {
-		this.role = role;
-		this.permission = permission;
-	}
-}
