+++ core/core-domain/src/main/java/com/pgms/coredomain/domain/member/enums/Role.java
@@ -1,0 +1,7 @@
+package com.pgms.coredomain.domain.member.enums;
+
+public enum Role {
+	ROLE_SUPERADMIN,
+	ROLE_ADMIN,
+	ROLE_USER
+}
