--- core/core-domain/src/main/java/com/pgms/coredomain/domain/member/repository/PermissionRepository.java
@@ -1,13 +1,0 @@
-package com.pgms.coredomain.domain.member.repository;
-
-import java.util.Optional;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import com.pgms.coredomain.domain.member.Permission;
-
-public interface PermissionRepository extends JpaRepository<Permission, Long> {
-	Optional<Permission> findByName(String name);
-
-	Optional<Permission> findByNameAndIdNot(String name, Long id);
-}
