--- core/core-domain/src/main/java/com/pgms/coredomain/domain/member/repository/RolePermissionRepository.java
@@ -1,9 +1,0 @@
-package com.pgms.coredomain.domain.member.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import com.pgms.coredomain.domain.member.RolePermission;
-
-public interface RolePermissionRepository extends JpaRepository<RolePermission, Long> {
-	int countByPermissionId(Long id);
-}
