--- core/core-domain/src/main/java/com/pgms/coredomain/domain/member/repository/RoleRepository.java
@@ -1,13 +1,0 @@
-package com.pgms.coredomain.domain.member.repository;
-
-import java.util.Optional;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import com.pgms.coredomain.domain.member.Role;
-
-public interface RoleRepository extends JpaRepository<Role, Long> {
-	Optional<Role> findByName(String name);
-
-	Optional<Object> findByNameAndIdNot(String name, Long id);
-}
