+++ core/core-security/src/main/java/com/pgms/coresecurity/security/service/OAuth2AuthenticationSuccessHandler.java
@@ -13,10 +13,8 @@ import org.springframework.security.web.authentication.AuthenticationSuccessHand
 import org.springframework.stereotype.Component;
 
 import com.pgms.coredomain.domain.member.Member;
-import com.pgms.coredomain.domain.member.Role;
 import com.pgms.coredomain.domain.member.enums.Provider;
 import com.pgms.coredomain.domain.member.repository.MemberRepository;
-import com.pgms.coredomain.domain.member.repository.RoleRepository;
 import com.pgms.coresecurity.security.jwt.JwtTokenProvider;
 import com.pgms.coresecurity.security.util.HttpResponseUtil;
 
@@ -32,7 +30,6 @@ import lombok.extern.slf4j.Slf4j;
 public class OAuth2AuthenticationSuccessHandler implements AuthenticationSuccessHandler {
 
 	private final MemberRepository memberRepository;
-	private final RoleRepository roleRepository;
 	private final JwtTokenProvider jwtTokenProvider;
 
 	@Override
@@ -58,18 +55,18 @@ public class OAuth2AuthenticationSuccessHandler implements AuthenticationSuccess
 		// 토큰 생성 후 반환
 		Map<String, Object> body = new HashMap<>();
 		body.put("accessToken", jwtTokenProvider.generateAccessToken((UserDetailsImpl)authenticated.getPrincipal()));
+		body.put("refreshToken", jwtTokenProvider.generateRefreshToken());
+
+		// TODO redis에 토큰 정보 저장
 		HttpResponseUtil.setSuccessResponse(response, HttpStatus.OK, body);
 	}
 
 	private Member createOrUpdateMember(String email, String name) {
-		return memberRepository.findByEmailWithRole(email).orElseGet(() -> {
-			Role defaultRole = roleRepository.findByName("ROLE_USER") //TODO: Role Enum으로 변경
-				.orElseThrow(() -> new RuntimeException("ROLE_USER가 존재하지 않습니다."));
+		return memberRepository.findByEmail(email).orElseGet(() -> {
 			Member newMember = Member.builder()
 				.email(email)
 				.name(name)
 				.provider(Provider.KAKAO)
-				.role(defaultRole)
 				.build();
 			return memberRepository.save(newMember);
 		});
